better visualization so that every rectangle keeps having the same color

+++ test/OptimalBinGeneratorTest.java
@@ -23,6 +23,6 @@ public class OptimalBinGeneratorTest {
                 bin.parameters.rectangles) {
             area += (rectangle.width * rectangle.height);
         }
-        Assertions.assertEquals(bin.optimal, area);
+        Assertions.assertEquals((int)bin.optimal, area);
     }
 }
